fix(cluster): accept human-readable units for --block-size and --bcast-chunk
Both flags were declared as plain usize without value_parser, so clap
parsed them as raw integers only. '16MiB' and '1GiB' were rejected with
'invalid digit found in string', forcing users to compute and pass raw
byte counts (16777216, 1073741824) - exactly what the v0.4.2 throttling
docs recommended *not* to do.

Promote both fields to u64 (matching --shard-size) and wire them to the
existing parse_size value parser. u64 also adds a layer of defense at the
parse boundary against the i32 truncation class of bug fixed in d38acad
(v0.3.1): a future narrow-int cast can't silently alias the chunk size
because the source type is u64 throughout.

usize::try_from at the single in-memory use site (broadcast.rs:123)
surfaces any 32-bit overflow with a clear error instead of a silent
truncation.

End-to-end CLI smoke test asserts that --block-size 16MiB --bcast-chunk
1GiB parses successfully, locking in the documented invocation form.

+++ crates/azcp-cluster/src/main.rs
@@ -166,7 +166,7 @@ fn main() -> ExitCode {
     // Resolve effective shard_size: explicit --shard-size wins; else
     // derive from --file-shards over the largest downloaded file; else 0
     // (one shard per file, v0.3.1 plan shape).
-    let bcast_chunk_u64 = args.bcast_chunk as u64;
+    let bcast_chunk_u64 = args.bcast_chunk;
     let effective_shard_size: u64 = if args.shard_size > 0 {
         args.shard_size
     } else if let Some(n) = args.file_shards {
